Refactoring
- Split asset and license blueprints and views into their own modules.
- Remove base /assets path
